% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{mylm}
\alias{mylm}
\title{mylm}
\usage{
mylm(formula, data)
}
\arguments{
\item{formula}{an object of class 'formula': a symbolic description of the model to be fitted.}

\item{data}{an optional data frame}
}
\value{
mylm returns an object of class 'lm'
}
\description{
Fit a linear model.
}
\examples{
mylm(mpg~cyl+disp, mtcars)

}
