% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm.R
\name{mylm}
\alias{mylm}
\title{mylm}
\usage{
mylm(f, data, penalty = 0, tol = 1e-17, useSparse = FALSE, useCpp = FALSE)
}
\arguments{
\item{data}{an optional data frame}

\item{formula}{an object of class 'formula': a symbolic description of the model to be fitted.}
}
\value{
mylm returns an object of class 'lm'
}
\description{
Fit a linear model.
}
\examples{
library(Rcpp)
library(plyr)
sourceCpp("sparseRregressionCPP.cpp")
mylm(mpg~cyl+disp, mtcars)

}
